feat(CAP-4): custom booking form btn
