+++ src/components/BookingForm.jsx
@@ -1,9 +1,19 @@
-import React from 'react'
+import { Col, Row } from "antd/lib/index";
+import React from "react";
 
 const BookingForm = () => {
   return (
-    <div>BookingForm</div>
-  )
-}
+    <div className="overflow-auto">
+      <Row className="px-2 py-4">
+        <Col xs={24} sm={24} md={24} lg={12}>
+          <h1 className="font-bold">Complete your Day Pass Booking</h1>
+        </Col>
+        <Col xs={24} sm={24} md={24} lg={12}>
+          b
+        </Col>
+      </Row>
+    </div>
+  );
+};
 
-export default BookingForm
+export default BookingForm;
